tests: move reusable components from `test.test_importer` to `test.helper`
`ImportHelper` and `AutotagStub` are used in many tests.
Thus, they should reside in a module which is obviously used by multiple
tests.
