+++ test/plugins/test_importadded.py
@@ -17,7 +17,7 @@
 
 import os
 import unittest
-from test.test_importer import AutotagStub, ImportHelper
+from test.helper import AutotagStub, ImportHelper
 
 from beets import importer
 from beets.util import displayable_path, syspath
