+++ test/plugins/test_mbsubmit.py
@@ -14,8 +14,13 @@
 
 
 import unittest
-from test.helper import TestHelper, capture_stdout, control_stdin
-from test.test_importer import AutotagStub, ImportHelper
+from test.helper import (
+    AutotagStub,
+    ImportHelper,
+    TestHelper,
+    capture_stdout,
+    control_stdin,
+)
 from test.test_ui_importer import TerminalImportSessionSetup
